tests: Move serde test case utilities to test_util test module
