+++ src/lib.rs
@@ -1,7 +1,12 @@
+pub use bool::Bool;
 pub use version::Version;
+
+#[cfg(test)]
+mod test_util;
 
 pub mod response;
 pub mod scenario;
 
+mod bool;
 mod util;
 mod version;
