+++ src/scenario/tests.rs
@@ -1,101 +1,8 @@
-use std::fmt::Debug;
-
 use indoc::indoc;
 
+use crate::test_util::{serde_test, value_from_str, value_to_string};
+
 use super::*;
-
-struct TestCase<T> {
-    repr: &'static str,
-    val: T,
-}
-
-impl<T: Eq + Serialize + Deserialize<'static> + std::fmt::Debug> TestCase<T> {
-    fn check<FS, FD>(&self, serialize_fn: FS, deserialize_fn: FD)
-    where
-        FS: Fn(&T) -> String,
-        FD: Fn(&'static str) -> T,
-    {
-        assert_eq!(
-            self.val,
-            (deserialize_fn)(self.repr),
-            "Incorrect deserialized value from '{}' (left: expected, right: actual)",
-            self.repr
-        );
-        assert_eq!(
-            self.repr,
-            (serialize_fn)(&self.val),
-            "Incorrect serialized value from '{:?}' (left: expected, right: actual)",
-            self.val
-        );
-    }
-}
-
-fn struct_to_string<T: Serialize + Debug>(val: &T) -> String {
-    match rfc822_like::to_string(val) {
-        Ok(t) => t,
-        Err(e) => panic!("Error when serializing {val:?}: {e}"),
-    }
-}
-
-fn struct_from_str<'de, T: Deserialize<'de>>(s: &'de str) -> T {
-    match rfc822_like::from_str(s) {
-        Ok(t) => t,
-        Err(e) => panic!("Error when deserializing \"{s}\": {e}"),
-    }
-}
-
-fn value_to_string<T: Serialize + Debug>(val: &T) -> String {
-    #[derive(Serialize, Debug)]
-    struct Record<'a, V> {
-        xxx: &'a V,
-    }
-
-    struct_to_string(&Record { xxx: val }).trim()["xxx: ".len()..].to_string()
-}
-
-fn value_from_str<T: for<'de> Deserialize<'de>>(s: &str) -> T {
-    #[derive(Deserialize)]
-    struct Record<V> {
-        xxx: V,
-    }
-
-    struct_from_str::<Record<T>>(&format!("xxx: {s}")).xxx
-}
-
-macro_rules! serde_test {
-    ($name:ident: {$($repr:expr => $val:expr),+}) => {
-        serde_test!(
-            $name(
-            struct_to_string,
-            struct_from_str
-            ): {$($repr => $val),+}
-        );
-    };
-
-    ($name:ident($serialize_fn:expr, $deserialize_fn:expr): {$($repr:expr => $val:expr),+}) => {
-        serde_test!(@test
-            $name,
-            $serialize_fn,
-            $deserialize_fn, $($repr, $val)+;
-        );
-    };
-
-    (@test $name:ident, $serialize_fn:expr, $deserialize_fn:expr, $($repr:expr, $val:expr)+;) => {
-        #[test]
-        fn $name() {
-            $(
-            {
-                let repr = { $repr };
-                let val = { $val };
-                TestCase {
-                    repr,
-                    val,
-                }.check($serialize_fn, $deserialize_fn);
-            }
-            )+
-        }
-    };
-}
 
 serde_test! {
     request: {
